move getGithubUrlForDocs to separate function bc of base branch update
