+++ packages/fern-dashboard/src/app/[orgName]/(visual-editor)/editor/[docsUrl]/[branch]/[...slug]/layout.tsx
@@ -15,9 +15,9 @@ import AbstractDefaultDocs from "@fern-docs/components/theming/AbstractDefaultDo
 import { GlobalStyles } from "@fern-docs/components/theming/global-styles";
 import { DesktopSearchButton } from "@fern-docs/search-ui/components/desktop/desktop-search-button";
 
-import getDocsGithubUrl from "@/app/api/get-docs-github-url/handler";
 import { getCurrentSession } from "@/app/services/auth0/getCurrentSession";
 import { Auth0OrgName } from "@/app/services/auth0/types";
+import getDocsGithubUrl from "@/app/services/dal/github/getDocsGithubUrl";
 import { assertGithubAccessByUrl } from "@/app/services/dal/github/validators";
 import { assertUserHasOrganizationAccess } from "@/app/services/dal/organization";
 import { GitHubLoader } from "@/app/services/github/github-loader";
@@ -70,10 +70,15 @@ export default async function VisualEditorPreviewLayout({
   });
 
   // Validate GitHub access
-  const githubUrl = await getDocsGithubUrl({
+  const urlResult = await getDocsGithubUrl({
     url: docsUrl,
     token: session.accessToken,
   });
+  if (!urlResult.success) {
+    redirect(`/${orgName}/docs`);
+  }
+
+  const { githubUrl } = urlResult;
   await assertGithubAccessByUrl(session.user.sub, githubUrl);
 
   // TODO: createEditableDocsLoader should be called here once, and data passed to child pages (@...) rather than called in those places as well
